move shovel state machine from GameManager to PowerUpManager
Shovel was the only power-up with a multi-frame state machine (timer,
flash phase, original-tile snapshot) hosted directly on GameManager.
Every other power-up was a thin _apply_* handler that delegated.

PowerUpManager now owns shovel_timer, the original-tile list, and the
flash logic. apply_shovel() and the per-frame tick run inside its
existing update(dt). GameManager._apply_shovel becomes a one-line
delegate, and _tick_shovel(dt) drops out of GameManager.update.

Tests: shovel state machine now lives in test_power_up_manager.py
against PowerUpManager directly (no GameManager mock scaffolding).
GameManager test keeps a delegation smoke test.
