+++ tests/unit/managers/test_game_manager_powerups.py
@@ -1,21 +1,16 @@
 import pytest
 import pygame
 from unittest.mock import MagicMock
-from src.core.tile import BrickVariant
 from src.managers.game_manager import GameManager
 from src.utils.constants import (
     PowerUpType,
     HELMET_INVINCIBILITY_DURATION,
     CLOCK_FREEZE_DURATION,
-    SHOVEL_DURATION,
-    SHOVEL_WARNING_DURATION,
-    SHOVEL_FLASH_INTERVAL,
     TankType,
     EffectType,
     TILE_SIZE,
 )
 from src.states.game_state import GameState
-from src.core.tile import TileType
 
 
 class TestPowerUpEffects:
@@ -123,75 +118,21 @@ class TestClockEffect:
         assert game.freeze_timer == CLOCK_FREEZE_DURATION
 
 
-class TestShovelEffect:
+class TestShovelDelegation:
+    """Shovel state machine lives in PowerUpManager; GameManager just delegates."""
+
     @pytest.fixture
     def game(self):
         gm = MagicMock(spec=GameManager)
         gm._apply_power_up = GameManager._apply_power_up.__get__(gm)
         gm._apply_shovel = GameManager._apply_shovel.__get__(gm)
-        gm._tick_shovel = GameManager._tick_shovel.__get__(gm)
         gm.state = GameState.RUNNING
-        gm.freeze_timer = 0.0
-        gm.shovel_timer = 0.0
-        gm._shovel_original_tiles = []
-        gm._shovel_flash_timer = 0.0
-        gm._shovel_flash_showing_steel = True
         mock_player = MagicMock()
-        mock_player.lives = 3
         gm.player_manager = MagicMock()
         gm.player_manager.get_active_players.return_value = [mock_player]
-        mock_tiles = []
-        for _ in range(4):
-            t = MagicMock()
-            t.type = TileType.BRICK
-            t.brick_variant = BrickVariant.FULL
-            mock_tiles.append(t)
-        gm.map = MagicMock()
-        gm.map.get_base_surrounding_tiles.return_value = mock_tiles
-        gm.map.set_tile_type = MagicMock()
+        gm.power_up_manager = MagicMock()
         return gm
 
-    def test_shovel_fortifies_base(self, game):
+    def test_apply_shovel_delegates_to_power_up_manager(self, game):
         game._apply_power_up(PowerUpType.SHOVEL)
-        assert game.shovel_timer == SHOVEL_DURATION
-        calls = game.map.set_tile_type.call_args_list
-        for call in calls:
-            assert call.args[1] == TileType.STEEL
-
-    def test_shovel_stores_originals(self, game):
-        game._apply_power_up(PowerUpType.SHOVEL)
-        assert len(game._shovel_original_tiles) == 4
-        for tile, orig_type in game._shovel_original_tiles:
-            assert orig_type == TileType.BRICK
-
-    def test_shovel_reverts_after_duration(self, game):
-        game._apply_power_up(PowerUpType.SHOVEL)
-        game.map.set_tile_type.reset_mock()
-        game._tick_shovel(SHOVEL_DURATION + 0.1)
-        assert game.shovel_timer <= 0
-        calls = game.map.set_tile_type.call_args_list
-        for call in calls:
-            assert call.args[1] == TileType.BRICK
-
-    def test_shovel_recollection_resets_timer(self, game):
-        game._apply_power_up(PowerUpType.SHOVEL)
-        original_tiles = game._shovel_original_tiles
-        game.shovel_timer = 5.0
-        game.map.set_tile_type.reset_mock()
-        mock_player = game.player_manager.get_active_players()[0]
-        game._apply_shovel(mock_player)
-        assert game.shovel_timer == SHOVEL_DURATION
-        assert game._shovel_original_tiles is original_tiles
-        game.map.set_tile_type.assert_not_called()
-
-    def test_shovel_flashes_during_warning(self, game):
-        game._apply_power_up(PowerUpType.SHOVEL)
-        game.map.set_tile_type.reset_mock()
-        # Advance into warning phase
-        game._tick_shovel(SHOVEL_DURATION - SHOVEL_WARNING_DURATION + 0.5)
-        # Advance past one flash interval to trigger toggle
-        game._tick_shovel(SHOVEL_FLASH_INTERVAL + 0.01)
-        # Verify toggle happened: tiles set to original type (BRICK)
-        assert game._shovel_flash_showing_steel is False
-        last_call = game.map.set_tile_type.call_args_list[-1]
-        assert last_call.args[1] == TileType.BRICK
+        game.power_up_manager.apply_shovel.assert_called_once_with()
