+++ tests/unit/managers/test_power_up_manager.py
@@ -2,8 +2,15 @@ import pytest
 import pygame
 from unittest.mock import MagicMock
 from src.managers.power_up_manager import PowerUpManager
-from src.core.tile import TileType
-from src.utils.constants import PowerUpType, TILE_SIZE, POWERUP_TIMEOUT
+from src.core.tile import BrickVariant, TileType
+from src.utils.constants import (
+    POWERUP_TIMEOUT,
+    PowerUpType,
+    SHOVEL_DURATION,
+    SHOVEL_FLASH_INTERVAL,
+    SHOVEL_WARNING_DURATION,
+    TILE_SIZE,
+)
 
 
 class TestPowerUpManager:
@@ -121,3 +128,59 @@ class TestPowerUpManager:
         player.rect = pygame.Rect(0, 0, TILE_SIZE, TILE_SIZE)
         manager.spawn_power_up(player, [])
         assert len(manager.active_power_ups) == 0
+
+
+class TestShovelEffect:
+    @pytest.fixture
+    def manager(self, mock_texture_manager):
+        game_map = MagicMock()
+        mock_tiles = []
+        for _ in range(4):
+            t = MagicMock()
+            t.type = TileType.BRICK
+            t.brick_variant = BrickVariant.FULL
+            mock_tiles.append(t)
+        game_map.get_base_surrounding_tiles.return_value = mock_tiles
+        return PowerUpManager(mock_texture_manager, game_map)
+
+    def test_initial_shovel_state(self, manager):
+        assert manager.shovel_timer == 0.0
+
+    def test_shovel_fortifies_base(self, manager):
+        manager.apply_shovel()
+        assert manager.shovel_timer == SHOVEL_DURATION
+        for call in manager._game_map.set_tile_type.call_args_list:
+            assert call.args[1] == TileType.STEEL
+
+    def test_shovel_stores_originals(self, manager):
+        manager.apply_shovel()
+        assert len(manager._shovel_original_tiles) == 4
+        for _, orig_type in manager._shovel_original_tiles:
+            assert orig_type == TileType.BRICK
+
+    def test_shovel_reverts_after_duration(self, manager):
+        manager.apply_shovel()
+        manager._game_map.set_tile_type.reset_mock()
+        manager.update(SHOVEL_DURATION + 0.1)
+        assert manager.shovel_timer <= 0
+        for call in manager._game_map.set_tile_type.call_args_list:
+            assert call.args[1] == TileType.BRICK
+
+    def test_shovel_recollection_resets_timer(self, manager):
+        manager.apply_shovel()
+        original_tiles = manager._shovel_original_tiles
+        manager.shovel_timer = 5.0
+        manager._game_map.set_tile_type.reset_mock()
+        manager.apply_shovel()
+        assert manager.shovel_timer == SHOVEL_DURATION
+        assert manager._shovel_original_tiles is original_tiles
+        manager._game_map.set_tile_type.assert_not_called()
+
+    def test_shovel_flashes_during_warning(self, manager):
+        manager.apply_shovel()
+        manager._game_map.set_tile_type.reset_mock()
+        manager.update(SHOVEL_DURATION - SHOVEL_WARNING_DURATION + 0.5)
+        manager.update(SHOVEL_FLASH_INTERVAL + 0.01)
+        assert manager._shovel_flash_showing_steel is False
+        last_call = manager._game_map.set_tile_type.call_args_list[-1]
+        assert last_call.args[1] == TileType.BRICK
